tipster-sim-full/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
matplotlib==3.9.0
numpy==1.26.4
tabulate==0.9.0
streamlit==1.36.0
